Add details on installing pyenv to manage python
When following the Docker tutorial it required me to use Python,
which is on the mac by default, but also **pip** to install
dependencies. That wasn't available.

So to keep things inline with how I have other languages setup I went
looking for a Python version manager and found
[Pyenv](https://github.com/pyenv/pyenv) to be the only player in town.

So this add details on installing **pyenv**, then a version of Python
and setting it as the global default.

For reference having followed this process nothing more was needed to
get [pip](https://en.wikipedia.org/wiki/Pip_(package_manager)) as it
comes with all versions of Python since 2.9.