Add a page to cover install and use of S3cmd tools
A number of our digital services make use of AWS S3 buckets, plus its typically used as a mechanism for transfering files between production environments and the developers.

Because I wanted to capture the basic commands I use, as well as how to install it I've added a new page to cover this tool.