Expand details on setting up iTerm
Have decided to switch to using iTerm2 over the default terminal on a Mac. With that decision there are some additional changes I make to get it setup how I like, hence it needs to move from a note on how to install it to a separate tool guide in its own right.