Update vim details to use solarized
Having settled on Solarized as my default theme and getting both Terminal and iTerm2 to use it, this also updates my vim configuration to make use of it.

It also includes adding details of a shortcut (F5) which when using Vim if pressed will switch the theme between solarized dark and solarized light.