Update post update to Catalina
This follows a wiper-and-fresh update to OSX Catalina.

When a major OSX upgrade comes about I take the opportunity to wipe and refresh my Mac's rather than just upgrade. When I did it this time round I found there were a number of things in my notes that were either outdated or irrelevant. Plus Catalina changes the default shell to `zsh` from `bash` so this meant a major update to how terminal, iterm and everything I do in them gets setup.