Update rbenv notes
Main updates are around how to con figure rbenv with zsh rather than bash.

Also added an entry on how to avoid errors when building older versions of ruby with the current homebrew version of ruby-build, and a reminder that Avira anitvirus is the main culprit when a version does not build!