Add alias helpers
With switching to using zsh and the oh-my-zsh plugins it comes with a number of aliases. This means I no longer need to manage my own. The one issue is that the aliases don't quite match up to what I was using.

So have added some alias helpers that allow me to quickly see what aliases are available for a given situation.