Update jenv and java notes
So the notes have been updated following the switch to zsh and oh-my-zsh. But also it had been some time since I set up Java and that has changed a great deal, especially how you go about installing it, and what versions of the JDK you can use.

Hence this is a big update to the Jenv entry.