Update post update to Catalina (#57)
This follows a wipe-and-refresh update to OSX Catalina.

When a major OSX upgrade comes about I take the opportunity to wipe and refresh my Mac's rather than just upgrade. When I did it this time around I found there were a number of things in my notes that were either outdated or irrelevant. Plus Catalina changes the default shell to `zsh` from `bash` so this meant a major update to how terminal, iTerm and everything I do in them gets setup.