Add new git alias to reset author
I had an issue with my homebrew installed version of nvm and using our `npm version patch` command in the WATER repos. It would work in the Docker container but fail locally.

So, my work around was to do it in the docker container and then reset the author of the commit generated to be me. Hopefully, now my nvm install is updated this shouldn't be an issue. But I created this alias and recording it here just in case.