Add pyenv and details for setting up python
Have started doing some work with python so capturing how I got things setup and a working environment running.

For reference, I got to this after a mismatch of Googling. [Setup your Python projects like a pro with Pyenv and Virtualenv on MacOS](https://medium.com/ml-hobbyist/setup-your-python-projects-like-a-pro-with-pyenv-and-virtualenv-on-macos-b4048b4d0f20) got me over the hump with virtual environments and [Managing Multiple Python Versions With pyenv](https://realpython.com/intro-to-pyenv/) confirmed what I'd done was ok!