Add TickTick to the base installations list
I've been using [Obsidian](https://obsidian.md/) to manage my notes and TODO list for some time. But a move to a managed Mac meant iCloud syncing was no longer possible.

A colleague suggested a plugin you could add to **Obsidian** which would sync it with a private GitHub repo. It was at that point I realised I didn't need a dedicated tool for keeping notes in markdown, I could do it all in VSCode and a private GitHub repo!

The one thing that would become tiresome when managing my notes in this way is my TODO list. Hence the search for a solution, and the decision to go with **TickTick** for now.